Make visitor a nested class.
