+++ pinot-query-runtime/src/test/java/org/apache/pinot/query/service/server/QueryServerTest.java
@@ -41,7 +41,7 @@ import org.apache.pinot.query.planner.physical.DispatchablePlanFragment;
 import org.apache.pinot.query.planner.physical.DispatchableSubPlan;
 import org.apache.pinot.query.planner.plannode.AbstractPlanNode;
 import org.apache.pinot.query.planner.plannode.PlanNode;
-import org.apache.pinot.query.planner.serde.PlanNodeSerializationVisitor;
+import org.apache.pinot.query.planner.serde.PlanNodeSerializer;
 import org.apache.pinot.query.routing.QueryPlanSerDeUtils;
 import org.apache.pinot.query.routing.QueryServerInstance;
 import org.apache.pinot.query.routing.StageMetadata;
@@ -227,7 +227,7 @@ public class QueryServerTest extends QueryTestSet {
   private Worker.QueryRequest getQueryRequest(DispatchableSubPlan queryPlan, int stageId) {
     DispatchablePlanFragment stagePlan = queryPlan.getQueryStageList().get(stageId);
     Plan.StageNode rootNode =
-        PlanNodeSerializationVisitor.INSTANCE.process((AbstractPlanNode) stagePlan.getPlanFragment().getFragmentRoot());
+        PlanNodeSerializer.process((AbstractPlanNode) stagePlan.getPlanFragment().getFragmentRoot());
     List<Worker.WorkerMetadata> workerMetadataList =
         QueryPlanSerDeUtils.toProtoWorkerMetadataList(stagePlan.getWorkerMetadataList());
     ByteString customProperty = QueryPlanSerDeUtils.toProtoProperties(stagePlan.getCustomProperties());
